fix(asd): asd

+++ src/features/comment/actions/delete-comment.ts
@@ -6,6 +6,7 @@ import {
   toFormState,
 } from "@/components/form/utils/to-form-state";
 import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
 import { ticketPath } from "@/paths";
 
@@ -13,10 +14,19 @@ export const deleteComment = async (id: string) => {
   const { user } = await getAuthOrRedirect();
 
   try {
+    const comment = await prisma.comment.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!comment || !isOwner(user, comment)) {
+      return toFormState("ERROR", "Not authorized");
+    }
+
     await prisma.comment.delete({
       where: {
         id,
-        userId: user.id,
       },
     });
   } catch (error) {
